Fix npc damage

+++ src/player/player_animations3.c
@@ -6,7 +6,7 @@
 /*   By: ohakola <ohakola@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/30 21:28:46 by ohakola           #+#    #+#             */
-/*   Updated: 2021/04/30 22:13:46 by ohakola          ###   ########.fr       */
+/*   Updated: 2021/04/30 22:18:37 by ohakola          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,13 +16,13 @@ static t_surface	*get_1080p_animation_source(t_doom3d *app)
 {
 	if (app->player.equipped_weapon->id == weapon_shotgun)
 		return ((t_surface*)hash_map_get(app->active_scene->animation_textures,
-			(int)"assets/animations/shotgun_anim_1080p.bmp"));
+			(long int)"assets/animations/shotgun_anim_1080p.bmp"));
 	else if (app->player.equipped_weapon->id == weapon_pistol)
 		return ((t_surface*)hash_map_get(app->active_scene->animation_textures,
-			(int)"assets/animations/pistol_anim_1080p.bmp"));
+			(long int)"assets/animations/pistol_anim_1080p.bmp"));
 	else if (app->player.equipped_weapon->id == weapon_fist)
 		return ((t_surface*)hash_map_get(app->active_scene->animation_textures,
-			(int)"assets/animations/fist_anim_1080p.bmp"));
+			(long int)"assets/animations/fist_anim_1080p.bmp"));
 	else if (app->player.equipped_weapon->id == weapon_rpg)
 	{
 		return ((t_surface*)hash_map_get(app->active_scene->animation_textures,
@@ -35,17 +35,17 @@ static t_surface	*get_720p_animation_source(t_doom3d *app)
 {
 	if (app->player.equipped_weapon->id == weapon_shotgun)
 		return ((t_surface*)hash_map_get(app->active_scene->animation_textures,
-			(int)"assets/animations/shotgun_anim_720p.bmp"));
+			(long int)"assets/animations/shotgun_anim_720p.bmp"));
 	else if (app->player.equipped_weapon->id == weapon_pistol)
 		return ((t_surface*)hash_map_get(app->active_scene->animation_textures,
-			(int)"assets/animations/pistol_anim_720p.bmp"));
+			(long int)"assets/animations/pistol_anim_720p.bmp"));
 	else if (app->player.equipped_weapon->id == weapon_fist)
 		return ((t_surface*)hash_map_get(app->active_scene->animation_textures,
-			(int)"assets/animations/fist_anim_720p.bmp"));
+			(long int)"assets/animations/fist_anim_720p.bmp"));
 	else if (app->player.equipped_weapon->id == weapon_rpg)
 	{
 		return ((t_surface*)hash_map_get(app->active_scene->animation_textures,
-			(int)"assets/animations/rpg_anim_720p.bmp"));
+			(long int)"assets/animations/rpg_anim_720p.bmp"));
 	}
 	return (NULL);
 }
@@ -54,17 +54,17 @@ static t_surface	*get_540p_animation_source(t_doom3d *app)
 {
 	if (app->player.equipped_weapon->id == weapon_shotgun)
 		return ((t_surface*)hash_map_get(app->active_scene->animation_textures,
-			(int)"assets/animations/shotgun_anim_540p.bmp"));
+			(long int)"assets/animations/shotgun_anim_540p.bmp"));
 	else if (app->player.equipped_weapon->id == weapon_pistol)
 		return ((t_surface*)hash_map_get(app->active_scene->animation_textures,
-			(int)"assets/animations/pistol_anim_540p.bmp"));
+			(long int)"assets/animations/pistol_anim_540p.bmp"));
 	else if (app->player.equipped_weapon->id == weapon_fist)
 		return ((t_surface*)hash_map_get(app->active_scene->animation_textures,
-			(int)"assets/animations/fist_anim_540p.bmp"));
+			(long int)"assets/animations/fist_anim_540p.bmp"));
 	else if (app->player.equipped_weapon->id == weapon_rpg)
 	{
 		return ((t_surface*)hash_map_get(app->active_scene->animation_textures,
-			(int)"assets/animations/rpg_anim_540p.bmp"));
+			(long int)"assets/animations/rpg_anim_540p.bmp"));
 	}
 	return (NULL);
 }
